clean up and re-organizing categories and folders
