+++ test/EventStore.Client.Streams.Tests/ReadStream/read_stream_forward.cs
@@ -1,8 +1,8 @@
 namespace EventStore.Client.Streams.Tests; 
 
-[Trait("Category", "Network")]
-[Trait("Category", "Stream")]
-[Trait("Category", "Read")]
+[Trait("Category", "Target:Stream")]
+[Trait("Category", "Operation:Read")]
+[Trait("Category", "Operation:Read:Forwards")]
 public class read_stream_forward(ITestOutputHelper output, EventStoreFixture fixture) : EventStoreTests<EventStoreFixture>(output, fixture) {
 	[Theory]
 	[InlineData(0)]
