+++ test/EventStore.Client.Streams.Tests/Subscriptions/SubscriptionsFixture.cs
@@ -1,6 +1,6 @@
 namespace EventStore.Client.Streams.Tests;
 
-[Trait("Category", "LongRunning")]
+
 [Trait("Category", "Subscriptions")]
 public class SubscriptionsFixture : EventStoreFixture {
 	public SubscriptionsFixture(): base(x => x.RunProjections()) {
