+++ test/EventStore.Client.Streams.Tests/Subscriptions/subscribe_to_stream.cs
@@ -1,6 +1,7 @@
 namespace EventStore.Client.Streams.Tests;
 
-[Trait("Category", "Stream")]
+[Trait("Category", "Subscriptions")]
+[Trait("Category", "Target:Stream")]
 public class subscribe_to_stream(ITestOutputHelper output, SubscriptionsFixture fixture) : EventStoreTests<SubscriptionsFixture>(output, fixture) {
 	[Fact]
 	public async Task receives_all_events_from_start() {
